Flight Booking System
Flight Booking System was created by me when i was in grade 10 in high school.
A project was to be made where i had to apply all the necessary concepts which i had grabbed.
The project is quite simple.

The manager (one who books tickets for users) has to login into the system. The system will ask where the user wants to fly, International or Domestic ?
(For domestic the country is India). The user choses any one. The system would ask to select the departure and arrival cities from the list displayed.
User details must be filled by the manager and in the end the system shows the total amount to be payed.

That's it !

The project is quite very basic and this project was a stepping stone into the programming world.

Signed-off by: Idrees Dargahwala